+++ rancilio-pid/controls.ino
@@ -4,10 +4,51 @@
  *****************************************************/
 #include "controls.h"
 #include "rancilio-debug.h"
-#include "rancilio-pid.h"
-//#include "userConfig.h"
+//#include "rancilio-pid.h"
 #include "MQTT.h"
+#include "blynk.h"
 #include "PIDBias.h"
+
+//TODO loaded from rancilio-pid.cpp. needs refactoring (class, injection)
+extern float* activeSetPoint;
+extern float* activeStartTemp;
+extern float* activeBrewTime;
+extern float* activePreinfusion;
+extern float* activePreinfusionPause;
+extern int pidON;
+extern char debugLine[200];
+extern unsigned int profile;
+extern float setPointSteam;
+
+//TODO old block to be removed by refactoring
+extern unsigned long userActivity;
+extern controlMap* controlsConfig;
+extern menuMap* menuConfig;
+extern const int OnlyPID;
+extern const int brewDetection;
+extern int brewing;
+extern void setGpioAction(int action, bool mode);
+extern int pumpRelayON, pumpRelayOFF;
+extern int valveRelayON, valveRelayOFF;
+extern int steaming;
+extern int cleaning;
+extern PIDBias bPID;
+extern unsigned long userActivitySavedOnForcedSleeping;
+extern int sleeping;
+extern unsigned long lastBrewEnd;
+extern unsigned int brewStatisticsAdditionalDisplayTime;
+extern bool MaschineColdstartRunOnce;
+extern float steadyPowerOffsetModified;
+extern unsigned int menuPosition;
+extern unsigned long previousTimerMenuCheck;
+extern unsigned int* activeBrewTimeEndDetection;
+extern float* activeScaleSensorWeightSetPoint;
+extern unsigned long brewTimer;
+extern float currentWeight;
+extern float steadyPowerOffset;
+extern unsigned long eepromForceSync;
+
+extern void blynkSave(char*);
 
 int simulatedBrewSwitch = 0;
 unsigned long previousCheckControls = 0;
