+++ rancilio-pid/eeprom-pcpid.ino
@@ -7,6 +7,8 @@
 #include "rancilio-debug.h"
 #include "rancilio-pid.h"
 
+extern void blynkSave(char*);
+
 void sync_eeprom() { sync_eeprom(false, false); }
 
 #ifdef ESP32
