--- rancilio-pid/helper.cpp
@@ -1,8 +1,0 @@
-#include "helper.h"
-
-int signnum(float x) {
-  if (x >= 0.0)
-    return 1;
-  else
-    return -1;
-}
