--- rancilio-pid/helper.h
@@ -1,6 +1,0 @@
-#ifndef HELPER_H
-#define HELPER_H
-
-int signnum(float x);
-
-#endif // HELPER_H
